New thing Lerning

+++ app/Http/Controllers/QuestionsController.php
@@ -8,6 +8,9 @@ use App\Http\Requests\AskQuestionRequest;
 
 class QuestionsController extends Controller
 {
+    public function __construct(){
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
     /**
      * Display a listing of the resource.
      *
@@ -63,6 +66,10 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
+        if(\Gate::denies('update-question', $question)){
+            abort(403, "Access Denaid");
+        }
+
         return view("questions.edit", compact('question'));
     }
 
@@ -75,6 +82,10 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        if(\Gate::denies('update-question', $question)){
+            abort(403, "Access Denaid");
+        }
+
         $question->update($request->only('title', 'body'));
         return redirect()->route('questions.index')->with('success', "Your Question Has Been Updated");
     }
@@ -87,6 +98,9 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+        if(\Gate::denies('delete-question', $question)){
+            abort(403, "Access Denaid");
+        }
         $question->delete();
         return redirect()->route('questions.index')->with('success', "Your Question Has Been Deleted");
     }
